Removed E-mail Notification
So the "not exitedCleanly" variable isn't working properly. Probably due
to multithreading.

And since these lines of code were purely for demonstration purposes
anyway I've commented them out so that we can easily reference and put
the e-mail sending back in later or those lines of code can simply be
deleted.