Fixed bug where cannot start without internet access
Fixed the bug where if you couldn't log a statistic to parse the
application would throw an exception and exit. Now the exception will be
caught, an error message printed, and then continue happily.
I didn't double check SysTrayIcon but it should work ^_^
(I don't have my LeapMotion on me to verify that it is running)